+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.2
+- added product custom fields to product export
+- fixed property values to be concatenated with a comma instead of being separate values
+
 # 0.1.1
 - added more support for FinSearch subscribers